introduce a module to house the REST code
