+++ build.gradle
@@ -114,6 +114,7 @@ subprojects {
       ':distribution:tools:keystore-cli',
       ':distribution:tools:launchers',
       ':distribution:tools:plugin-cli',
+      ':modules:rest-compatibility',
       ':qa:os',
       ':qa:wildfly',
       ':x-pack:plugin:autoscaling',
