+++ server/src/main/java/org/elasticsearch/action/ActionModule.java
@@ -641,7 +641,6 @@ public class ActionModule extends AbstractModule {
         registerHandler.accept(new RestIndexAction.CompatibleCreateHandler());
         registerHandler.accept(new RestIndexAction.CompatibleAutoIdHandler(nodesInCluster));
         registerHandler.accept(new RestGetAction());
-        registerHandler.accept(new RestGetAction.CompatibleRestGetAction());
         registerHandler.accept(new RestGetSourceAction());
         registerHandler.accept(new RestMultiGetAction(settings));
         registerHandler.accept(new RestDeleteAction());
